refactor: move testUtils
