+++ src/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import path from 'path'
-import { cases } from '../testUtils'
+import { cases } from '../utils/testUtils'
 
 const filename = path.resolve(__dirname, '..', '__fixtures__', 'messages.js')
